cnn_paddle_final/requirements.txt
imgaug==0.4.0
matplotlib==2.0.2
numpy==1.13.1
opencv_python==4.10.0.84
paddlepaddle==2.6.2
pandas==2.2.3
Pillow==4.2.1
scikit_learn==1.3.2
seaborn==0.13.2
tqdm==4.66.4

